statemanager uses push/pop fns
this means that going to the pause state and then coming back won't
result in reinitializing the playstate. structure of statemanager will
support in long term, ability to switch between many different states.
each state gets left in the state it was previously in, until it becomes
the current state.